+++ jksoa-common/src/main/kotlin/net/jkcode/jksoa/common/_Logger.kt
@@ -3,8 +3,6 @@ package net.jkcode.jksoa.common
 import net.jkcode.jkutil.common.ModuleLogSwitcher
 
 internal val switcher = ModuleLogSwitcher("rpc")
-// 注册中心的日志
-val registerLogger = switcher.getLogger("net.jkcode.jksoa.rpc.registry")
 // 客户端的日志
 val clientLogger = switcher.getLogger("net.jkcode.jksoa.rpc.client")
 val connLogger = switcher.getLogger("net.jkcode.jksoa.rpc.client.connection")