+++ jksoa-rpc/jksoa-rpc-client/src/main/kotlin/net/jkcode/jksoa/rpc/client/IReferer.kt
@@ -4,8 +4,6 @@ import net.jkcode.jksoa.common.loader.IServiceClass
 
 /**
  * 服务的引用（代理）
- *   1 引用服务
- *   2 向注册中心订阅服务
  *
  * @Description:
  * @author shijianhang<772910474@qq.com>