+++ jksoa-rpc/jksoa-rpc-client/src/main/kotlin/net/jkcode/jksoa/rpc/client/k8s/server/PatternServerResolver.kt
@@ -5,7 +5,7 @@ import net.jkcode.jksoa.rpc.client.k8s.K8sUtil
 
 /**
  * 根据模式来解析的k8s server解析器，有缓存
- *   从rpc请求(rpc服务类)中,解析出k8s服务名(server:协议ip端口)
+ *   从rpc请求(rpc服务类)中,解析出k8s应用域名(server:协议ip端口)
  *
  * @author shijianhang<772910474@qq.com>
  * @date 2022-5-9 3:18 PM
@@ -13,7 +13,7 @@ import net.jkcode.jksoa.rpc.client.k8s.K8sUtil
 object PatternServerResolver : IServerResolver {
 
     /**
-     * 包名转为k8s服务名(server)的映射配置
+     * 包名转为k8s应用域名(server)的映射配置
      * key是包名的模式: 用.分割多层包名, *代表一层任意包名, **代表多层任意包名
      * value是server地址: 可以带变量, 变量格式为`$层序号`, 如$0代表第1层包名, $1代表第2层包名, 以此类推
      */
@@ -35,7 +35,7 @@ object PatternServerResolver : IServerResolver {
     private val resolveCache: MutableMap<String, String> = HashMap()
 
     /**
-     * 解析k8s服务名(server)
+     * 解析k8s应用域名(server)
      * @param serviceId
      * @return 协议ip端口(server)
      */