+++ jksoa-rpc/jksoa-rpc-client/src/main/kotlin/net/jkcode/jksoa/rpc/client/referer/Referer.kt
@@ -7,8 +7,6 @@ import net.jkcode.jksoa.rpc.client.IReferer
 
 /**
  * 服务的引用（代理）
- *   1 引用服务
- *   2 向注册中心订阅服务
  *
  * @Description:
  * @author shijianhang<772910474@qq.com>