+++ jksoa-rpc/jksoa-rpc-server/src/main/kotlin/net/jkcode/jksoa/rpc/server/IProvider.kt
@@ -5,8 +5,6 @@ import net.jkcode.jksoa.common.Url
 
 /**
  * 服务提供者
- *   1 提供服务
- *   2 向注册中心注册服务
  *
  * @Description:
  * @author shijianhang<772910474@qq.com>